requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
reportlab==4.2.5
openai==1.58.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
